Add Address / Primeface / show address

+++ src/de/oio/jsf/Address.java
@@ -4,7 +4,6 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 
 @Entity
@@ -12,7 +11,7 @@ public class Address {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int id;
+	private Long id;
 	private String street;
 	private int zip;
 	private String city;
@@ -36,11 +35,11 @@ public class Address {
 		this.person = person;
 	}
 
-	public int getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 
@@ -67,5 +66,6 @@ public class Address {
 	public void setCity(String city) {
 		this.city = city;
 	}
+	
 
 }
